fix(bulkUploadWorkersMappings): Update mappings import in lambdas spec and rename countries to country
